DL-17828 | Change Identifier type from Int to Long

+++ app/uk/gov/hmrc/pla/stub/model/hip/HIPProtectionsModel.scala
@@ -53,7 +53,7 @@ object ProtectionRecordsList {
 }
 
 case class ProtectionRecord(
-    identifier: Int,
+    identifier: Long,
     sequenceNumber: Int,
     `type`: ProtectionType,
     certificateDate: String,
